refactor table
