+++ benchmark/FedHPOB/fedhpob/benchmarks/tabular_benchmark.py
@@ -15,8 +15,6 @@ class TabularBenchmark(BaseBenchmark):
                  rng=None,
                  cost_mode='estimated',
                  **kwargs):
-        self.model, self.dname, self.algo, self.cost_mode = model, dname, \
-                                                            algo, cost_mode
         self.table, self.meta_info = load_data(datadir, model, dname, algo)
         self.eval_freq = self.meta_info['eval_freq']
         super(TabularBenchmark, self).__init__(model, dname, algo, cost_mode,
@@ -30,24 +28,6 @@ class TabularBenchmark(BaseBenchmark):
         for key, value in fidelity.items():
             assert value in self.fidelity_space[
                 key], 'fidelity invalid, check `fidelity_space` for help.'
-
-    def _search(self, configuration, fidelity):
-        # For configuration
-        mask = np.array([True] * self.table.shape[0])
-        for col in configuration.keys():
-            mask *= (self.table[col].values == configuration[col])
-        idx = np.where(mask)
-        result = self.table.iloc[idx]
-
-        # For fidelity
-        mask = np.array([True] * result.shape[0])
-        for col in fidelity.keys():
-            if col == 'round':
-                continue
-            mask *= (result[col].values == fidelity[col])
-        idx = np.where(mask)
-        result = result.iloc[idx]["result"]
-        return result
 
     def objective_function(self,
                            configuration,
